# public/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Stubborn
  description: "Documentation générée automatiquement pour l'application Stubborn."
  contact:
    name: 'Support Stubborn'
    email: contact@lewebpluschouette.fr
  license:
    name: MIT
    url: 'https://opensource.org/licenses/MIT'
  version: 1.0.0
paths:
...
  '/cart/add/{id}':
    post:
      tags:
        - Panier
      summary: 'Ajouter un article au panier'
      operationId: cartAdd
      parameters:
        -
          name: id
          in: path
          description: 'ID de la variante de produit'
          required: true
          schema:
            type: integer
      requestBody:
        description: 'Quantité (form-data ou JSON). Défaut : 1'
        required: false
        content:
          application/json:
            schema:
              properties:
                qty:
                  type: integer
                  default: 1
                  minimum: 1
              type: object
      responses:
        '302':
          description: 'Redirection vers le panier'
        '404':
          description: 'Produit introuvable'
  '/cart/remove/{id}':
    post:
      tags:
        - Panier
      summary: 'Retirer un article du panier'
      operationId: cartRemove
      parameters:
        -
          name: id
          in: path
          description: 'ID de la variante à retirer'
          required: true
          schema:
            type: integer
      responses:
        '302':
          description: 'Redirection après suppression'
        '404':
          description: 'Article introuvable'
...
  /checkout:
    post:
      tags:
        - Paiement
      summary: 'Créer la session Stripe et rediriger'
      operationId: checkoutStartPost
      responses:
        '303':
          description: 'See Other → URL Stripe'
        '401':
          description: 'Non authentifié'
        '400':
          description: 'Panier vide'
      security:
        -
          cookieAuth: []
  '/checkout/success/{id}':
    get:
      tags:
        - Paiement
      summary: 'Afficher la page de confirmation de commande'
      operationId: checkoutSuccess
      parameters:
        -
          name: id
          in: path
          description: 'Identifiant de la commande'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Page succès avec détails commande'
        '404':
          description: 'Commande introuvable'
...
  /products:
    get:
      tags:
        - Produits
      summary: 'Lister les produits'
      operationId: listProducts
      parameters:
        -
          name: range
          in: query
          description: 'Filtrer par fourchette de prix (optionnel)'
          required: false
          schema:
            type: string
            enum:
              - 10-29
              - 30-35
              - 36-50
      responses:
        '200':
          description: 'Page HTML ou JSON de produits'
  '/product/{id}':
    get:
      tags:
        - Produits
      summary: 'Afficher un produit'
      operationId: getProduct
      parameters:
        -
          name: id
          in: path
          description: 'Identifiant du produit'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Page détail produit'
        '404':
          description: 'Produit non trouvé'
  /admin:
    get:
      tags:
        - Produits
      summary: 'Tableau de bord admin des produits'
      operationId: adminDashboard
      responses:
        '200':
          description: 'Vue HTML du dashboard'
    post:
      tags:
        - Produits
      summary: 'Créer un produit (admin)'
      operationId: createProduct
      responses:
        '302':
          description: 'Redirection après création'
        '400':
          description: 'Données invalides'
  '/admin/delete/{id}':
    post:
      tags:
        - Produits
      summary: 'Supprimer un produit'
      operationId: deleteProduct
      parameters:
        -
          name: id
          in: path
          description: 'ID du produit à supprimer'
          required: true
          schema:
            type: integer
      responses:
        '302':
          description: 'Redirection après suppression'
        '404':
          description: 'Produit introuvable'
  '/admin/product/{id}/edit':
    get:
      tags:
        - Produits
      summary: "Afficher le formulaire d'édition (admin)"
      operationId: editProductForm
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Formulaire HTML'
    post:
      tags:
        - Produits
      summary: 'Mettre à jour un produit'
      operationId: updateProduct
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '302':
          description: 'Redirection après mise à jour'
        '400':
          description: 'Données invalides'
...